\name{bmmProject-package}
\alias{bmmProject-package}
\alias{bmmProject}
\docType{package}
\title{
\packageTitle{bmmProject}
}
\description{
\packageDescription{bmmProject}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{bmmProject}

Maintainer: \packageMaintainer{bmmProject}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = bmmProject)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{bmmProject}}
%\section{Documentation Index}{\packageIndices{bmmProject}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
